Fixing problems in the ISO parsing, and testsuite ...
(1) The [clock scan] heuristics for handling a missing month and/or
    day are in conflict with the ISO specification. When encountering
    such a pattern we now rewrite pattern and input to force the ISO
    defaults.

(2) Fixed the regular expression for the time information in a full
    timestamp as it did not handle that : in a TZ offset is optional,
    nor that minutes and seconds are optional (The code after the
    MatchTime was prepared for such missing fields however).

(2a) Again missing HMS fields got the wrong defaults. We are now
     rewriting the pattern and input to force the proper ISO defaults
     (start of day, hour, minute).

(3) Fixed the tests which thought that 199707 as YYYYMM. ISO
    (Wikipedia) explicitly says that this is YYMMDD instead.

(4) Fixed the test numbering, had some duplicates.

(5) Fixed one test with wrong results, expected 11:20 for input
    specifying 11:00.